% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_hyppar.R
\name{assign_hyppar}
\alias{assign_hyppar}
\title{Assign Hyperparameters Based on Data Density}
\usage{
assign_hyppar(data, ddt = dt_density_tbl)
}
\arguments{
\item{data}{data.table. contains columns for location_id and assigned
densities}

\item{ddt}{data density table, will use dt_density_tbl by default}
}
\value{
data table containing locations, data density value and
corresponding hyperparameters for use in space-time gpr
}
\description{
Assign hyperparameters based on data density scores
for use in space-time GPR. Density scores can be produced by
calc_data_density(). hyperparameters are assigned based on
level of data density defined by a data density table
(see dt_density_tbl)
}
\examples{
data <- calc_data_density(Belgium, "vr", source_type = "source_type")
data <- assign_hyppar(data)
}
