% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Belgium}
\alias{Belgium}
\title{Belgium}
\format{
A data.table with 18 rows and 4 variables:
\describe{
\item{location_id}{unique id per location}
\item{year_id}{year}
\item{source_id}{unique id per data source}
\item{source_type}{type of source}
}
}
\source{
not important
}
\usage{
Belgium
}
\description{
A dataset containing years and sources of SRB data in Belgium 2005-2022
}
\keyword{datasets}
