% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_density.R
\name{calc_data_density}
\alias{calc_data_density}
\title{Calculate Data Density}
\usage{
calc_data_density(
  data,
  vars,
  weights = 1,
  source_type = "source_type",
  source_id = "nid",
  location_id = "location_id",
  year_id = "year_id"
)
}
\arguments{
\item{data}{data.table which should contain the following columns
\itemize{
\item \emph{source_type}: data collection source e.g. contains ("vr", "cbh", "census")
\item \emph{location_id}: like ihme_loc_id.
\item \emph{year}: years of focus
\item \emph{source_id}: unique id for data sources
unique location-source-years will count towards data density. data density is
calculated per location.
}}

\item{vars}{character. types of source which will be added up to calculate data density}

\item{weights}{vector of numerics. each corresponds with a source_type in vars}

\item{source_type}{character, name of column which tells type of source}

\item{source_id}{column name for nid}

\item{location_id}{unique id for location}

\item{year_id}{column name detailing years of study}
}
\value{
data table with locations and corresponding data density scores
}
\description{
Calculate Data Density
}
\examples{
data <- calc_data_density(sample_reports,
"source_type", "location_id", c("vr", "cbh"), c(1, .5))
check that:
source_type:year_id are strings which are column names in data
vars is type char
weights is numeric and either length 1 or the same length as vars
}
