% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_density_tbl}
\alias{dt_density_tbl}
\title{Data Density Table}
\format{
A data.table with 5 rows and 9 variables:
\describe{
\item{data_density_category}{category name for hyperparameters}
\item{density_min & density_max}{bounds for data_density. interval is
min inclusive}
\item{hyperparameters}{parameters for use in ST-GPR. to learn more read
Process Documentation on the HUB}
...
}
}
\source{
check the HUB
}
\usage{
dt_density_tbl
}
\description{
A dataset containing data density categories used in
hyperparameter assignment for ST-GPR
}
\keyword{datasets}
