% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters_srb.R
\name{hyperparameters_srb}
\alias{hyperparameters_srb}
\title{TODO:
theoretical spacetime prep function: likely missing steps, but it is a proof
of concept}
\usage{
hyperparameters_srb(data, ddt = dt_density_tbl)
}
\description{
Steps for full process:
0. modify ddt (optional)
This Function:
\enumerate{
\item calcs data density in way in method
\item assigns hyperparams based on ddt
}
}
