% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_densities.R
\name{mod_densities}
\alias{mod_densities}
\title{Modify Data Density Categories}
\usage{
mod_densities(cutpoints, ddt = dt_density_tbl)
}
\arguments{
\item{cutpoints}{numeric vector defining data density categories.
lowest value should be 0, largest value defines an open interval}

\item{ddt}{data table containing data densitys and hyperparams.
will use dt_density_tbl if none is provided}
}
\value{
data density table with updated data density categories.
}
\description{
Modify Data Density Categories
}
\examples{
new_ddt <- mod_densities(c(0,15,30,40,50))
}
